# platform/frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "DemandRestartRequest": {
        "additionalProperties": false,
        "properties": {
          "expected_generation": {
            "maximum": 9007199254740991.0,
            "minimum": 0.0,
            "title": "Expected Generation",
            "type": "integer"
          },
          "expected_sequence": {
            "maximum": 9007199254740991.0,
            "minimum": 0.0,
            "title": "Expected Sequence",
            "type": "integer"
          },
          "idempotency_key": {
            "maxLength": 200,
            "minLength": 1,
            "pattern": "\\S",
            "title": "Idempotency Key",
            "type": "string"
          },
          "rationale": {
            "maxLength": 2000,
            "minLength": 1,
            "pattern": "\\S",
            "title": "Rationale",
            "type": "string"
          }
        },
        "required": [
          "idempotency_key",
          "expected_generation",
          "expected_sequence",
          "rationale"
        ],
        "title": "DemandRestartRequest",
        "type": "object"
      },
      "DemandRestartResponse": {
        "additionalProperties": false,
        "properties": {
          "change_sequence": {
            "title": "Change Sequence",
            "type": "integer"
          },
          "demand_id": {
            "title": "Demand Id",
            "type": "string"
          },
          "generation": {
            "title": "Generation",
            "type": "integer"
          },
          "occurrence_id": {
            "title": "Occurrence Id",
            "type": "string"
          },
          "restart_id": {
            "title": "Restart Id",
            "type": "string"
          },
          "state": {
            "const": "preparing",
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "restart_id",
          "demand_id",
          "occurrence_id",
          "generation",
          "change_sequence",
          "state"
        ],
        "title": "DemandRestartResponse",
        "type": "object"
      },
      "DemandStatusResponse": {
        "additionalProperties": false,
        "properties": {
          "change_sequence": {
            "default": 0,
            "maximum": 9007199254740991.0,
            "minimum": 0.0,
            "title": "Change Sequence",
            "type": "integer"
          },
          "current_run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Run Id"
          },
          "demand_id": {
            "title": "Demand Id",
            "type": "string"
          },
          "generation": {
            "maximum": 9007199254740991.0,
            "minimum": 0.0,
            "title": "Generation",
            "type": "integer"
          },
          "not_before": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Not Before"
          },
          "occurrence_id": {
            "title": "Occurrence Id",
            "type": "string"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "retry_attempt": {
            "maximum": 9007199254740991.0,
            "minimum": 0.0,
            "title": "Retry Attempt",
            "type": "integer"
          },
          "run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          },
          "state": {
            "enum": [
              "preparing",
              "coalescing",
              "queued",
              "running",
              "updated",
              "retained",
              "needs_review",
              "retry_wait",
              "retry_exhausted",
              "cannot_recompute",
              "paused"
            ],
            "title": "State",
            "type": "string"
          },
          "withdrawn_basis_pair_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Withdrawn Basis Pair Ids"
          }
        },
        "required": [
          "demand_id",
          "occurrence_id",
          "state",
          "generation",
          "retry_attempt",
          "run_id",
          "reason",
          "not_before"
        ],
        "title": "DemandStatusResponse",
        "type": "object"
      },
      "ErrorDetailResponse": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "additionalProperties": true,
            "title": "Details",
            "type": "object"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "details"
        ],
        "title": "ErrorDetailResponse",
        "type": "object"
      },
      "ErrorEnvelopeResponse": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetailResponse"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelopeResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Anonymous trusted-intranet crash analysis control plane. There are intentionally no login, RBAC, or DELETE endpoints.",
    "title": "Crash-Cap API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v3/workspaces/{workspace_id}/occurrences/{occurrence_id}/analysis-demand": {
      "get": {
        "operationId": "get_analysis_demand_api_v3_workspaces__workspace_id__occurrences__occurrence_id__analysis_demand_get",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "title": "Workspace Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "occurrence_id",
            "required": true,
            "schema": {
              "title": "Occurrence Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/DemandStatusResponse"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Get Analysis Demand Api V3 Workspaces  Workspace Id  Occurrences  Occurrence Id  Analysis Demand Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          }
        },
        "summary": "Get Analysis Demand"
      }
    },
    "/api/v3/workspaces/{workspace_id}/occurrences/{occurrence_id}/analysis-demand/restarts": {
      "post": {
        "operationId": "restart_analysis_demand_api_v3_workspaces__workspace_id__occurrences__occurrence_id__analysis_demand_restarts_post",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "title": "Workspace Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "occurrence_id",
            "required": true,
            "schema": {
              "title": "Occurrence Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DemandRestartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DemandRestartResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeResponse"
                }
              }
            },
            "description": "Crash-Cap error envelope"
          }
        },
        "summary": "Restart Analysis Demand"
      }
    },
...
  }
}